+++ src/test/index.ts
@@ -1,3 +1,4 @@
+import * as fs from 'fs';
 import * as glob from 'glob';
 import * as Mocha from 'mocha';
 import * as path from 'path';
@@ -6,15 +7,17 @@ import { Configurations } from '../constants';
 
 export async function run(): Promise<void> {
     // Create the mocha test
+
     const mocha = new Mocha({
         ui: 'bdd',
         color: true,
         inlineDiffs: true,
-        timeout: 7000,
+        timeout: Number.POSITIVE_INFINITY,
         slow: 5000,
+        bail: true,
     });
 
-    const testsRoot = path.resolve(__dirname, '..');
+    const testsRoot = path.resolve(__dirname);
     await workspace.getConfiguration('').update('window.restoreFullscreen', true, true);
     await workspace.getConfiguration('').update('window.newWindowDimensions', 'maximized', true);
     await workspace.getConfiguration('').update('workbench.editor.decorations.badges', false, true);
@@ -22,13 +25,17 @@ export async function run(): Promise<void> {
     await workspace.getConfiguration('').update(Configurations.SaveGlobally, false, true);
 
     return new Promise((c, e) => {
-        glob('./test/suite/**/**.test.js', { cwd: testsRoot }, (err, files) => {
+        const secondTime = fs.existsSync(path.resolve(testsRoot, '..', '..', '.vscode-test', 'folder', '12.txt'));
+        const globPattern = secondTime ? './suite/**/**.spec.js' : './suite/**/**.test.js';
+        glob(globPattern, { cwd: testsRoot }, (err, files) => {
             if (err) {
                 return e(err);
             }
 
             // Add files to the test suite
-            files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+            files.forEach(f => {
+                mocha.addFile(path.resolve(testsRoot, f));
+            });
 
             try {
                 // Run the mocha test