+++ src/test/suite/saveAndRestore.test.ts
@@ -2,8 +2,8 @@ import { expect } from 'chai';
 import * as robot from 'kbm-robot';
 import * as path from 'path';
 import { commands, Uri, window, workspace } from 'vscode';
-import { BuiltInCommands, Commands } from '../../constants';
-import { delay, expectEditors, newGroupName } from '../utils';
+import { BuiltInCommands, Commands, serializeTabGroups } from '../../constants';
+import { delay, newGroupName } from '../utils';
 
 describe('Save and Restore', () => {
     let workspaceFolder: string;
@@ -21,9 +21,14 @@ describe('Save and Restore', () => {
         await commands.executeCommand(BuiltInCommands.CloseAllEditorGroups);
     });
 
+    afterEach(async () => {
+        // await window.showQuickPick([{ label: 'Yes' }], { placeHolder: 'Continue', ignoreFocusOut: true });
+    });
+
     it('Single file, default name', async function () {
         const fileuri = Uri.file(path.join(workspaceFolder, './1.txt'));
         await window.showTextDocument(fileuri, { preview: false });
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString('\n').go();
@@ -33,7 +38,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString('Group 1\n').go();
         await wait;
-        await expectEditors([fileuri]);
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('Two files, single editor group', async function () {
@@ -42,6 +52,7 @@ describe('Save and Restore', () => {
         const fileuri2 = Uri.file(path.join(workspaceFolder, './2.txt'));
         await window.showTextDocument(fileuri1, { preview: false });
         await window.showTextDocument(fileuri2, { preview: false });
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -51,7 +62,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1, fileuri2], { viewColumns: [1, 1] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('Two files, different editor groups', async function () {
@@ -60,6 +76,7 @@ describe('Save and Restore', () => {
         const fileuri2 = Uri.file(path.join(workspaceFolder, './2.txt'));
         await window.showTextDocument(fileuri1, { viewColumn: 1, preview: false });
         await window.showTextDocument(fileuri2, { viewColumn: 2, preview: false });
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -69,7 +86,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1, fileuri2], { viewColumns: [1, 2] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('One file, different editor groups', async function () {
@@ -78,6 +100,7 @@ describe('Save and Restore', () => {
         const fileuri2 = Uri.file(path.join(workspaceFolder, './1.txt'));
         await window.showTextDocument(fileuri1, { viewColumn: 1, preview: false });
         await window.showTextDocument(fileuri2, { viewColumn: 2, preview: false });
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -87,7 +110,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1, fileuri2], { viewColumns: [1, 2] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('Single pinned file', async function () {
@@ -95,6 +123,7 @@ describe('Save and Restore', () => {
         const fileuri1 = Uri.file(path.join(workspaceFolder, './1.txt'));
         await window.showTextDocument(fileuri1, { preview: false });
         await commands.executeCommand(BuiltInCommands.PinEditor);
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -104,7 +133,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1], { pins: [true] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('Two pinned files, same editor group', async function () {
@@ -115,6 +149,7 @@ describe('Save and Restore', () => {
         await commands.executeCommand(BuiltInCommands.PinEditor);
         await window.showTextDocument(fileuri2, { preview: false });
         await commands.executeCommand(BuiltInCommands.PinEditor);
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -124,7 +159,12 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1, fileuri2], { pins: [true, true] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
     it('Two pinned files, different editor group', async function () {
@@ -135,6 +175,7 @@ describe('Save and Restore', () => {
         await commands.executeCommand(BuiltInCommands.PinEditor);
         await window.showTextDocument(fileuri2, { viewColumn: 2, preview: false });
         await commands.executeCommand(BuiltInCommands.PinEditor);
+        const tabGroups = serializeTabGroups(window.tabGroups);
         let wait = commands.executeCommand(Commands.Save);
         await delay(100);
         await robot.typeString(groupName).go();
@@ -144,20 +185,25 @@ describe('Save and Restore', () => {
         await delay(100);
         await robot.typeString(groupName).go();
         await wait;
-        await expectEditors([fileuri1, fileuri2], { viewColumns: [1, 2], pins: [true, true] });
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 
-    it('10 * random different editor groups', async function () {
-        const tests = 10;
+    const tests = 3;
+    it.only(`${tests} * random different editor groups`, async function () {
         this.timeout(this.timeout() * tests);
         this.slow(this.slow() * tests);
         for (let i = 0; i < tests; i++) {
             await commands.executeCommand(BuiltInCommands.CloseAllEditorGroups);
             const groupName = newGroupName();
-            const numFiles = ((Math.trunc(Math.random() * 100)) % 7) + 3;
+            const numFiles = ((Math.trunc(Math.random() * 100)) % 4) + 10;
             const filenames = `${Math.random()}`.slice(2).split('').map(s => +s + 1).slice(0, numFiles);
             const fileUris = filenames.map(f => Uri.file(path.join(workspaceFolder, `${f}.txt`)));
-            const isNewViewColumn = filenames.map((_, j) => Math.random() < 0.3 || j === 0 ? true : false);
+            const isNewViewColumn = filenames.map((_, j) => Math.random() < 0.2 || j === 0 ? true : false);
             const viewColumns: number[] = [];
             let viewColumn = 0;
             let thisViewColumn: number[] = [];
@@ -177,6 +223,7 @@ describe('Save and Restore', () => {
                 await window.showTextDocument(f, { viewColumn: viewColumns[j], preview: false });
                 if (pins[j]) await commands.executeCommand(BuiltInCommands.PinEditor);
             }
+            const tabGroups = serializeTabGroups(window.tabGroups);
             let wait = commands.executeCommand(Commands.Save);
             await delay(100);
             await robot.typeString(groupName).go();
@@ -186,7 +233,48 @@ describe('Save and Restore', () => {
             await delay(100);
             await robot.typeString(groupName).go();
             await wait;
-            await expectEditors(fileUris, { viewColumns, pins });
+            const newTabGroups = serializeTabGroups(window.tabGroups);
+            if (tabGroups !== newTabGroups) {
+                expect.fail('Tab groups not restored');
+            }
+            // const error = checkTabGroups(tabGroups, window.tabGroups);
+            // if (error) expect.fail(error);
         }
+    });
+
+    it.skip('Specific set', async function () {
+        const auto = true;
+        this.timeout(Infinity);
+        const groupName = newGroupName();
+        await commands.executeCommand(BuiltInCommands.CloseAllEditorGroups);
+        const filenames = [2, 2, 6, 6, 3, 7, 9, 8, 2, 9, 6,];
+        const fileUris = filenames.map(f => Uri.file(path.join(workspaceFolder, `${f}.txt`)));
+        const viewColumns = [1, 2, 2, 3, 3, 3, 3, 3, 3, 4, 4];
+        const pins = [true, false, false, false, false, false, false, false, true, false, false,];
+        for (let j = 0; j < fileUris.length; j++) {
+            const f = fileUris[j];
+            await window.showTextDocument(f, { viewColumn: viewColumns[j], preview: false });
+            if (pins[j]) await commands.executeCommand(BuiltInCommands.PinEditor);
+        }
+        const tabGroups = serializeTabGroups(window.tabGroups);
+        let wait = commands.executeCommand(Commands.Save);
+        if (auto) {
+            await delay(100);
+            await robot.typeString(groupName).go();
+        }
+        await wait;
+        await commands.executeCommand(BuiltInCommands.CloseAllEditorGroups);
+        wait = commands.executeCommand(Commands.Restore);
+        if (auto) {
+            await delay(100);
+            await robot.typeString(groupName).go();
+        }
+        await wait;
+        const newTabGroups = serializeTabGroups(window.tabGroups);
+        if (tabGroups !== newTabGroups) {
+            expect.fail('Tab groups not restored');
+        }
+        // const error = checkTabGroups(tabGroups, window.tabGroups);
+        // if (error) expect.fail(error);
     });
 });